Fix ConnServer stall

+++ include/comm/Protocol.h
@@ -8,7 +8,14 @@
 
 namespace comm {
 
-    ENUM_FLAGS(Protocol, uint8_t)
+    // We are setting Protocol to be uint32_t here because
+    // otherwise there will be padding, and we
+    // observed that sometimes that padding is causing issues
+    // on the client side when decoding the protocol.
+    // TODO: Make HelloMessage and Protocol a Protobuf message.
+    // to prevent this kind of issues, and to prevent
+    // potential issues with endianness.
+    ENUM_FLAGS(Protocol, uint32_t)
     {
         None = 0,
         TCP = 1 << 0,
